+++ test/controllers/users_controller_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class UsersControllerTest < ActionDispatch::IntegrationTest
   def setup
-    @user = users :neil
+    @user = users :michael
     @other_user = users :archer
   end
 
@@ -49,5 +49,24 @@ class UsersControllerTest < ActionDispatch::IntegrationTest
     assert_not @other_user.admin?
     patch user_path(@other_user), params: { user: { password: 'password', password_confirmation: 'password', admin: true } }
     assert_not @other_user.reload.admin?
+  end
+
+  test 'should redirect destroy when logged in' do
+    assert_no_difference 'User.count' do
+      delete user_path(@user)
+    end
+
+    assert_response :see_other
+    assert_redirected_to login_url
+  end
+
+  test 'should redirect destroy when logged in as a non-admin' do
+    log_in_as @other_user
+    assert_no_difference 'User.count' do
+      delete user_path(@user)
+    end
+
+    assert_response :see_other
+    assert_redirected_to login_url
   end
 end